requirements.txt
scrapli_netconf
black
pyyaml
xmltodict